refactor(onboarding): deconstruct OnboardingGuideComponents.tsx into separate illustration files
