+++ src/components/organisms/OnboardingGuide.tsx
@@ -2,7 +2,7 @@ import { ChevronLeft, ChevronRight, Sparkles, X } from "lucide-react"
 import React, { useState } from "react"
 
 import { useLanguage } from "../../contexts/LanguageContext"
-import { getSteps } from "./OnboardingGuideComponents"
+import { getSteps } from "./OnboardingGuide/OnboardingSteps"
 
 interface OnboardingGuideProps {
   isOpen: boolean
